+++ packages/audit-log/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @appspine/audit-log
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @appspine/common@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes