+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @appspine/auth
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @appspine/common@0.1.1
+  - @appspine/audit-log@0.1.1
+
 ## 0.1.2
 
 ### Patch Changes