+++ packages/health-check/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @appspine/health-check
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @appspine/common@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes